Refactor the Media Upload to have a separate component for upload.
